Remove old classes in Market module
